add setup tab
